+++ prisma/migrations/20240115040451_init/migration.sql
@@ -18,20 +18,13 @@ CREATE TABLE "WeekDay" (
 -- CreateTable
 CREATE TABLE "Alarm" (
     "id" SERIAL NOT NULL,
+    "userId" INTEGER NOT NULL,
     "weekDayId" INTEGER NOT NULL,
+    "enabled" BOOLEAN NOT NULL DEFAULT true,
     "startTime" TEXT NOT NULL,
     "endTime" TEXT NOT NULL,
 
     CONSTRAINT "Alarm_pkey" PRIMARY KEY ("id")
-);
-
--- CreateTable
-CREATE TABLE "UserAlarm" (
-    "id" SERIAL NOT NULL,
-    "userId" INTEGER NOT NULL,
-    "alarmId" INTEGER NOT NULL,
-
-    CONSTRAINT "UserAlarm_pkey" PRIMARY KEY ("id")
 );
 
 -- CreateIndex
@@ -40,11 +33,11 @@ CREATE UNIQUE INDEX "User_username_key" ON "User"("username");
 -- CreateIndex
 CREATE UNIQUE INDEX "WeekDay_name_key" ON "WeekDay"("name");
 
+-- CreateIndex
+CREATE UNIQUE INDEX "Alarm_weekDayId_userId_key" ON "Alarm"("weekDayId", "userId");
+
 -- AddForeignKey
 ALTER TABLE "Alarm" ADD CONSTRAINT "Alarm_weekDayId_fkey" FOREIGN KEY ("weekDayId") REFERENCES "WeekDay"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
 
 -- AddForeignKey
-ALTER TABLE "UserAlarm" ADD CONSTRAINT "UserAlarm_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
-
--- AddForeignKey
-ALTER TABLE "UserAlarm" ADD CONSTRAINT "UserAlarm_alarmId_fkey" FOREIGN KEY ("alarmId") REFERENCES "Alarm"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+ALTER TABLE "Alarm" ADD CONSTRAINT "Alarm_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;