+++ src/routes/alarmsRouter.ts
@@ -1,25 +1,13 @@
 import { Router } from 'express';
 
+import alarmsController from '../controllers/alarmsController';
+import verifyToken from '../middlewares/verifyToken';
+
 const alarmsRouter = Router();
-
-const defaultAlarm = {
-  enabled: true,
-  startTime: '09:00',
-  endTime: '17:00',
-};
-
-const alarms = {
-  monday: defaultAlarm,
-  tuesday: defaultAlarm,
-  wednesday: defaultAlarm,
-  thursday: defaultAlarm,
-  friday: defaultAlarm,
-  saturday: defaultAlarm,
-  sunday: defaultAlarm,
-};
+alarmsRouter.use(verifyToken);
 
 alarmsRouter.get('/', (req, res) => {
-  res.status(200).json(alarms);
+  alarmsController.getAlarms(req, res);
 });
 
 export default alarmsRouter;